refactor(notification): mv balance threshold event definitions
